requirements.txt
fastapi 
uvicorn 
pydantic 
python-dotenv 
langchain-groq
groq
langchain-community
langchain-core 
httpx 
sqlalchemy 
psycopg2-binary 
redis 
langchain 
chromadb 
sentence-transformers 
reportlab 
whisper
pandas